+++ CHANGELOG.md
@@ -1,3 +1,11 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+- This is Initail release of Flutter package which can be used for loging and sending report.
+
+## 0.0.2
+
+- Minor change in 3rd party package version
+
+## 0.0.3
+
+- Add delete zip file in initialization method